tests/test_automated_trading.py
import pytest
from src.api.moomoo_api import MooMooAPI
from src.features.automated_trading import AutomatedTrading
from unittest.mock import Mock, patch

@pytest.fixture
def mock_moomoo_api():
    return Mock(spec=MooMooAPI)

@pytest.fixture
def mock_moomoo_openD():
    return Mock()

@pytest.fixture
def automated_trading(mock_moomoo_api, mock_moomoo_openD):
    return AutomatedTrading(mock_moomoo_api, mock_moomoo_openD)

def test_automated_trading_initialization(automated_trading):
    assert automated_trading is not None
    assert automated_trading.moomoo_api is not None
    assert automated_trading.moomoo_openD is not None

# Add more test cases as needed
