requirements.txt
requests
PyYAML
moomoo-api
TA-Lib
